requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
pandas>=2.0.0
torch>=2.0.0
